# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/vote": {
      "get": {
        "operationId": "VoteController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of votes to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of votes to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "postId",
            "required": false,
            "in": "query",
            "description": "Filter by post ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by user ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by vote type (upvote/downvote)",
            "schema": {
              "enum": [
                1,
                -1
              ],
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of votes with total count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "votes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Vote"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 100
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all votes with optional filters",
        "tags": [
          "votes"
        ]
      },
      "post": {
        "operationId": "VoteController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vote created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vote"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Post not found"
          },
          "409": {
            "description": "User has already voted on this post"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vote"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new vote",
        "tags": [
          "votes"
        ]
      }
    },
    "/vote/{id}": {
      "delete": {
        "operationId": "VoteController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vote removed successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to remove this vote"
          },
          "404": {
            "description": "Vote not found"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remove a vote",
        "tags": [
          "votes"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "fullname": {
            "type": "string",
            "description": "User full name"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string",
            "description": "User role"
          }
        },
        "required": [
          "_id",
          "email",
          "username",
          "password",
          "fullname",
          "role"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "Vote": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của lượt bình chọn"
          },
          "user": {
            "description": "Người dùng thực hiện bình chọn",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của người dùng thực hiện bình chọn"
          },
          "targetType": {
            "enum": [
              "post",
              "comment"
            ],
            "type": "string",
            "description": "Loại đối tượng được bình chọn (bài viết hoặc bình luận)",
            "example": "post"
          },
          "targetId": {
            "type": "string",
            "description": "ID của đối tượng được bình chọn"
          },
          "voteValue": {
            "enum": [
              1,
              -1
            ],
            "type": "number",
            "description": "Giá trị bình chọn (1 cho upvote, -1 cho downvote)",
            "example": 1
          },
          "post": {
            "description": "Bài viết được bình chọn (nếu có)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "comment": {
            "description": "Bình luận được bình chọn (nếu có)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Comment"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bình chọn"
          }
        },
        "required": [
          "_id",
          "user",
          "userId",
          "targetType",
          "targetId",
          "voteValue",
          "createdAt"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của bình luận"
          },
          "post": {
            "description": "Bài viết chứa bình luận này",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "postId": {
            "type": "string",
            "description": "ID của bài viết"
          },
          "user": {
            "description": "Tác giả bình luận",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của tác giả"
          },
          "content": {
            "type": "string",
            "description": "Nội dung bình luận"
          },
          "parentComment": {
            "description": "Bình luận cha (nếu là bình luận trả lời)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Comment"
              }
            ]
          },
          "parentCommentId": {
            "type": "string",
            "description": "ID của bình luận cha (nếu có)"
          },
          "replies": {
            "description": "Các bình luận trả lời",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "votes": {
            "description": "Các lượt bình chọn của bình luận",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bình luận"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm cập nhật bình luận"
          }
        },
        "required": [
          "_id",
          "post",
          "postId",
          "user",
          "userId",
          "content",
          "replies",
          "votes",
          "createdAt",
          "updatedAt"
        ]
      },
      "Post": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của bài viết"
          },
          "title": {
            "type": "string",
            "description": "Tiêu đề bài viết",
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "description": "Nội dung bài viết"
          },
          "user": {
            "description": "Tác giả bài viết",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của tác giả"
          },
          "category": {
            "description": "Danh mục của bài viết",
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "ID của danh mục"
          },
          "tags": {
            "description": "Các thẻ được gắn cho bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "comments": {
            "description": "Các bình luận của bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "votes": {
            "description": "Các lượt bình chọn của bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          },
          "isLocked": {
            "type": "boolean",
            "description": "Trạng thái khóa của bài viết"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bài viết"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm cập nhật bài viết"
          }
        },
        "required": [
          "_id",
          "title",
          "content",
          "user",
          "userId",
          "category",
          "categoryId",
          "tags",
          "comments",
          "votes",
          "isLocked",
          "createdAt",
          "updatedAt"
        ]
      },
      "Tag": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của thẻ"
          },
          "tagName": {
            "type": "string",
            "description": "Tên của thẻ",
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "description": "Mô tả của thẻ"
          },
          "posts": {
            "description": "Các bài viết được gắn thẻ này",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          }
        },
        "required": [
          "_id",
          "tagName",
          "posts"
        ]
      },
...
      "CreateVoteDto": {
        "type": "object",
        "properties": {
          "targetId": {
            "type": "string",
            "description": "ID of the target (post or comment) to vote on",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "targetType": {
            "enum": [
              "post",
              "comment"
            ],
            "type": "string",
            "description": "Type of the target (post or comment)",
            "example": "post"
          },
          "voteValue": {
            "enum": [
              1,
              -1
            ],
            "type": "number",
            "description": "Type of vote (upvote or downvote)",
            "example": 1
          }
        },
        "required": [
          "targetId",
          "targetType",
          "voteValue"
        ]
      },
...
    }
  }
}